Use QueryEntityOptions() for faster performance.
The read() function now uses QueryEntitiesOptions() to return only the 'data' property of the entity. No other properties are required and they increased the size of the response unnecessarily. 

The gc() function now uses QueryEntitiesOptions() to return only the RowKey of the entities. This is the only property required for the delete operation and all other properties being returned, especially the 'data' property, increased the size of the response considerably.